refactor: restructure app routes with route groups and simplify admin-check middleware
- Move all user-facing pages under app/(app)/(user)/ route group with a
  shared layout, separating them from the top-level (app) layout
- Move admin pages under app/(app)/(user)/(admin)/ with a dedicated admin
  layout for admin-only access control at the layout level
- Refactor checkAdminPermission() to delegate to isAdmin() utility instead
  of issuing a redundant direct DB query, removing ~10 lines of duplicated logic
- Update admin-check test description to reflect new behavior (no authError check)
