--- app/(app)/admin/users/page.tsx
@@ -1,88 +1,0 @@
-"use client";
-
-import { useEffect, useState } from "react";
-import { useRouter } from "next/navigation";
-import { useTranslations } from "next-intl";
-import { createClient } from "@/lib/supabase/client";
-import { AdminUserView, UserStatus } from "@/types/database";
-import { UserManagementList } from "@/components/admin/user-management-list";
-import { Loader2 } from "lucide-react";
-
-interface UserWithStats extends AdminUserView {
-  is_admin: boolean;
-  status: UserStatus;
-  stats: {
-    prediction_count: number;
-    tournament_count: number;
-    total_points: number;
-  };
-}
-
-export default function AdminUsersPage() {
-  const t = useTranslations("admin.users");
-  const [users, setUsers] = useState<UserWithStats[]>([]);
-  const [loading, setLoading] = useState(true);
-  const [isAdmin, setIsAdmin] = useState(false);
-  const router = useRouter();
-  const supabase = createClient();
-
-  useEffect(() => {
-    async function checkAuthAndLoadUsers() {
-      // Check if user is logged in
-      const {
-        data: { user },
-      } = await supabase.auth.getUser();
-
-      if (!user) {
-        router.push("/login");
-        return;
-      }
-
-      // Check if user is admin
-      const { data: userProfile } = await supabase
-        .from("users")
-        .select("is_admin")
-        .eq("id", user.id)
-        .single();
-
-      if (!userProfile?.is_admin) {
-        router.push("/unauthorized");
-        return;
-      }
-
-      setIsAdmin(true);
-
-      // Fetch users with stats
-      try {
-        const response = await fetch("/api/admin/users");
-        if (!response.ok) throw new Error("Failed to fetch users");
-        const data = await response.json();
-        setUsers(data);
-      } catch (error) {
-        console.error("Error loading users:", error);
-      } finally {
-        setLoading(false);
-      }
-    }
-
-    checkAuthAndLoadUsers();
-  }, [router, supabase]);
-
-  if (loading || !isAdmin) {
-    return (
-      <div className="container mx-auto py-8 px-4 flex items-center justify-center min-h-[400px]">
-        <Loader2 className="h-8 w-8 animate-spin text-muted-foreground" />
-      </div>
-    );
-  }
-
-  return (
-    <div className="container mx-auto py-8 px-4">
-      <div className="mb-8">
-        <h1 className="text-4xl font-bold mb-2">{t("title")}</h1>
-        <p className="text-muted-foreground">{t("subtitle")}</p>
-      </div>
-      <UserManagementList initialUsers={users} />
-    </div>
-  );
-}
