--- app/(app)/teams/new/page.tsx
@@ -1,47 +1,0 @@
-import { createClient } from "@/lib/supabase/server";
-import { redirect } from "next/navigation";
-import { requireAdmin } from "@/lib/utils/admin";
-import { TeamCreateForm } from "@/components/teams/management/team-create-form";
-import Link from "next/link";
-import { Button } from "@/components/ui/button";
-import { ArrowLeft } from "lucide-react";
-import { getTranslations } from "next-intl/server";
-
-export default async function NewTeamPage() {
-  const t = await getTranslations("teams");
-  const supabase = await createClient();
-
-  const {
-    data: { user },
-  } = await supabase.auth.getUser();
-
-  if (!user) {
-    redirect("/login");
-  }
-
-  try {
-    await requireAdmin();
-  } catch {
-    redirect("/unauthorized");
-  }
-
-  return (
-    <div className="container mx-auto py-8 px-4">
-      <div className="max-w-2xl mx-auto">
-        <Link href="/teams">
-          <Button variant="ghost" className="mb-6">
-            <ArrowLeft className="h-4 w-4 mr-2" />
-            {t("backToTeams")}
-          </Button>
-        </Link>
-
-        <div className="mb-8">
-          <h1 className="text-4xl font-bold mb-2">{t("create.title")}</h1>
-          <p className="text-muted-foreground">{t("create.subtitle")}</p>
-        </div>
-
-        <TeamCreateForm />
-      </div>
-    </div>
-  );
-}
