+++ lib/middleware/admin-check.ts
@@ -1,5 +1,6 @@
 import { NextResponse } from "next/server";
 import { createClient } from "@/lib/supabase/server";
+import { isAdmin } from "@/lib/utils/admin";
 
 /**
  * Middleware to check admin permissions in API routes
@@ -21,20 +22,14 @@ export async function checkAdminPermission(): Promise<NextResponse | null> {
   const supabase = await createClient();
   const {
     data: { user },
-    error: authError,
   } = await supabase.auth.getUser();
 
-  if (authError || !user) {
+  if (!user) {
     return NextResponse.json({ error: "Unauthorized: Authentication required" }, { status: 401 });
   }
 
-  const { data: userProfile } = await supabase
-    .from("users")
-    .select("is_admin")
-    .eq("id", user.id)
-    .single();
-
-  if (!userProfile?.is_admin) {
+  const admin = await isAdmin();
+  if (!admin) {
     return NextResponse.json({ error: "Forbidden: Admin access required" }, { status: 403 });
   }
